refactor(output): split XML formatter sections
Keep the XML formatter entry point focused on document flow while moving document-level and page-level XML section rendering into focused modules. This preserves the output shape while making the formatter easier to review and extend.
